Decoupled business logic and ui
